# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP System API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parties:
    get:
      operationId: listParties
      tags: [parties]
      summary: List parties (customers/suppliers)
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [customer, supplier, both]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of parties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartyListResponse"
    post:
      operationId: createParty
      tags: [parties]
      summary: Create a party
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartyBody"
      responses:
        "201":
          description: Party created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
  /parties/{id}:
    get:
      operationId: getParty
      tags: [parties]
      summary: Get party details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Party details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
    patch:
      operationId: updateParty
      tags: [parties]
      summary: Update party
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartyBody"
      responses:
        "200":
          description: Party updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
    delete:
      operationId: deleteParty
      tags: [parties]
      summary: Delete party
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Party deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /parties/{id}/ledger:
    get:
      operationId: getPartyLedger
      tags: [parties]
      summary: Get party ledger (statement)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Party ledger entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartyLedgerResponse"

  # ITEMS
...
  /units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List units of measurement
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Unit"
                required: [data]
    post:
      operationId: createUnit
      tags: [units]
      summary: Create a unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUnitBody"
      responses:
        "201":
          description: Unit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{id}:
    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Delete unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unit deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # HSN CODES
  /hsn:
    get:
      operationId: listHsnCodes
      tags: [hsn]
      summary: List HSN codes
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of HSN codes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/HsnCode"
                required: [data]
    post:
      operationId: createHsnCode
      tags: [hsn]
      summary: Create HSN code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHsnCodeBody"
      responses:
        "201":
          description: HSN code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HsnCode"

  # TAX RATES
  /tax-rates:
    get:
      operationId: listTaxRates
      tags: [tax-rates]
      summary: List tax rates
      responses:
        "200":
          description: List of tax rates
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaxRate"
                required: [data]
    post:
      operationId: createTaxRate
      tags: [tax-rates]
      summary: Create tax rate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaxRateBody"
      responses:
        "201":
          description: Tax rate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxRate"

  # CUSTOM FIELDS
  /custom-fields:
    get:
      operationId: listCustomFields
      tags: [custom-fields]
      summary: List custom field definitions
      parameters:
        - name: entity
          in: query
          schema:
            type: string
            enum: [item, party, voucher, voucher_item]
      responses:
        "200":
          description: List of custom fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CustomField"
                required: [data]
    post:
      operationId: createCustomField
      tags: [custom-fields]
      summary: Create custom field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomFieldBody"
      responses:
        "201":
          description: Custom field created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomField"
  /custom-fields/{id}:
    delete:
      operationId: deleteCustomField
      tags: [custom-fields]
      summary: Delete custom field
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Custom field deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # SALES INVOICES
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Party:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [customer, supplier, both]
        gstin:
          type: string
        pan:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        stateCode:
          type: string
        pincode:
          type: string
        openingBalance:
          type: number
        openingBalanceType:
          type: string
          enum: [debit, credit]
        creditLimit:
          type: number
        creditDays:
          type: integer
        isActive:
          type: boolean
        currentBalance:
          type: number
        customFields:
          type: object
        createdAt:
          type: string
      required: [id, name, type, isActive]
    PartyListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Party"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total]
    CreatePartyBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [customer, supplier, both]
        gstin:
          type: string
        pan:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        stateCode:
          type: string
        pincode:
          type: string
        openingBalance:
          type: number
        openingBalanceType:
          type: string
          enum: [debit, credit]
        creditLimit:
          type: number
        creditDays:
          type: integer
        customFields:
          type: object
      required: [name, type]
    PartyLedgerEntry:
      type: object
      properties:
        date:
          type: string
        voucherType:
          type: string
        voucherNumber:
          type: string
        description:
          type: string
        debit:
          type: number
        credit:
          type: number
        balance:
          type: number
      required: [date, voucherType, voucherNumber, debit, credit, balance]
    PartyLedgerResponse:
      type: object
      properties:
        party:
          $ref: "#/components/schemas/Party"
        openingBalance:
          type: number
        closingBalance:
          type: number
        entries:
          type: array
          items:
            $ref: "#/components/schemas/PartyLedgerEntry"
      required: [openingBalance, closingBalance, entries]
...
    Unit:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        symbol:
          type: string
      required: [id, name, symbol]
    CreateUnitBody:
      type: object
      properties:
        name:
          type: string
        symbol:
          type: string
      required: [name, symbol]
    HsnCode:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: string
        taxRate:
          type: number
      required: [id, code]
    CreateHsnCodeBody:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        taxRate:
          type: number
      required: [code]
    TaxRate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        rate:
          type: number
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
      required: [id, name, rate]
    CreateTaxRateBody:
      type: object
      properties:
        name:
          type: string
        rate:
          type: number
      required: [name, rate]
    CustomField:
      type: object
      properties:
        id:
          type: integer
        entity:
          type: string
          enum: [item, party, voucher, voucher_item]
        fieldName:
          type: string
        fieldLabel:
          type: string
        fieldType:
          type: string
          enum: [text, number, date, boolean, select]
        options:
          type: array
          items:
            type: string
        isRequired:
          type: boolean
        sortOrder:
          type: integer
      required: [id, entity, fieldName, fieldLabel, fieldType]
    CreateCustomFieldBody:
      type: object
      properties:
        entity:
          type: string
          enum: [item, party, voucher, voucher_item]
        fieldName:
          type: string
        fieldLabel:
          type: string
        fieldType:
          type: string
          enum: [text, number, date, boolean, select]
        options:
          type: array
          items:
            type: string
        isRequired:
          type: boolean
        sortOrder:
          type: integer
      required: [entity, fieldName, fieldLabel, fieldType]
...